\name{HDF5Array-internals}

\alias{HDF5Array-internals}
\alias{shorten_assay2h5_links}
\alias{restore_absolute_assay2h5_links}
\alias{validate_HDF5ArraySeed_dataset}
\alias{write_h5_assays}
\alias{create_dir}
\alias{replace_dir}
\alias{check_and_delete_files}
\alias{stop_if_bad_dir}

\title{HDF5Array internals}

\description{
    Internal utilities defined in the \pkg{HDF5Array} package. These functions
    are not intended to be used directly.
}

\keyword{internal}
